api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.6.0
anthropic>=0.40.0
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
pydantic>=2.0.0
